调整代码结构
todo：调整函数和方法存放位置以及命名
