+++ app/factories/__init__.py
@@ -1,0 +1,4 @@
+from .device_factory import DeviceFactory
+from .function_factory import FunctionFactory
+from .link_factory import LinkFactory
+from .operation_factory import OperationFactory
