+++ app/factories/base.py
@@ -1,0 +1,3 @@
+class Factory:
+    def generate(self):
+        pass
