+++ app/solver/__init__.py
