+++ contrib/example_stack.py
@@ -1,0 +1,3 @@
+from app.configure_process import processor
+
+processor("./mock/input_1")
