+++ uniform_model/actions/group.py
@@ -1,0 +1,5 @@
+from uniform_model import DeviceGroup
+
+
+def group(*devices):
+    return DeviceGroup(*devices)
