+++ uniform_model/devices/__init__.py
@@ -1,0 +1,3 @@
+from .device import Device
+from .device_manager import DeviceManager
+from .group import DeviceGroup
