+++ uniform_model/functions/ip.py
@@ -1,0 +1,15 @@
+import logging
+
+from uniform_model.functions.base import Function
+
+logging.basicConfig(format='%(asctime)s <%(name)s> %(message)s')
+logger = logging.getLogger('uniform_model.functions.ip')
+logger.setLevel(logging.DEBUG)
+
+
+class FunctionIP(Function):
+    attrs = ['bootproto', 'broadcast', 'ipaddr', 'netmask', 'network']
+
+    @classmethod
+    def op(cls, arith_list):
+        print(f'op on {arith_list}')
