+++ uniform_model/functions/stack.py
@@ -1,75 +1,10 @@
 import logging
-from collections import Counter
-from functools import reduce
-from models.interfaces import Interface
+
+from uniform_model.functions.base import Function
 
 logging.basicConfig(format='%(asctime)s <%(name)s> %(message)s')
-logger = logging.getLogger('functions')
+logger = logging.getLogger('uniform_model.functions.stack')
 logger.setLevel(logging.DEBUG)
-
-
-class OperableTrait:
-    @classmethod
-    def op(cls, *arith_list, **kwargs): raise NotImplemented()
-
-
-class FunctionType(type):
-    def __str__(cls):
-        return f'<Protocol: {cls.__name__}>'
-
-
-class Function(OperableTrait, metaclass=FunctionType):
-    dependencies = []
-
-    @classmethod
-    def validate(cls, device):
-        """
-        自定义协议和模型检查流程，在接口检查后自动调用
-        :return: 检查通过返回真，否则 raise对应错误(推荐) 或者 返回False
-        """
-        try:
-            if not cls._check_protocol_support(device):
-                raise ProtocolNotSupport(f"device {device.id} not support FunctionStack")
-            if not cls._check_dependencies(device):
-                return False
-        except ProtocolNotSupport as e:
-            print(e)
-            return False
-        return True
-
-    @classmethod
-    def _check_protocol_support(cls, device):
-        return cls.name in device.support_functions
-
-    @classmethod
-    def _check_dependencies(cls, device):
-        for dependency in cls.dependencies:
-            if not getattr(device, dependency).enable:
-                return False
-        return True
-
-    def generate_conf(self):
-        pass
-
-
-class FunctionIP(Function):
-    attrs = ['bootproto', 'broadcast', 'ipaddr', 'netmask', 'network']
-
-    @classmethod
-    def op(cls, arith_list):
-        print(f'op on {arith_list}')
-
-
-class FunctionVLAN(Function):
-    @classmethod
-    def op(cls, arith_list):
-        print(f'op on {arith_list}')
-
-
-class FunctionTrunk(Function):
-    @classmethod
-    def op(cls, arith_list):
-        print(f'op on {arith_list}')
 
 
 class FunctionStack(Function):
@@ -174,7 +109,3 @@ class FunctionStack(Function):
             output.append("commit")
             output.append("quit")
         return output
-
-
-class ProtocolNotSupport(Exception):
-    pass
