+++ uniform_model/templates/__init__.py
@@ -1,0 +1,4 @@
+from .base import Template
+from .template_manager import TemplateManager
+from .chassis import ChassisTemplate
+from .switch import SwitchTemplate
